refactor: extract shared indicator formatting and category constants
Consolidate duplicated logic across indicator components:
- Extract CATEGORY_ORDER/CATEGORY_LABELS to indicator-registry
- Extract formatIndicatorField() shared helper
- Replace CARD_HELP dict with descriptor.description
